web/requirements.txt
Django==5.2.3
numpy==2.1.3
opencv-python==4.11.0.86
pillow==11.2.1
tensorflow==2.19.0
